Participant editable
Participants can now be edited by admin. In case a user submits wrong
data ("Jane Doe -> Male") this is crucial. Wrong data can occur also if
Gender-Toggle is not working correctly (occured on iPhones). In case of
gender change feedback is removed from the RatingQuestion corresponding
to the old gender and then added to the correct RatingQuestion.